Project page is divided into two other components
